// src/model/user.ts
export interface user { 
    iduser: number;
    name: string;
    lastName: string;
    documentType: number;
    documentTypeName: string;
    document: string;
    phone: string;
    indicativeId: number;
    departamentId: number;
    cityId: number;
    address: string;
    dateBirth: Date;
    email: string;
    userName: string;
    password: string;
}